refactor(services): request, response type 모듈화
