+++ src/services/rankingProductList/types.ts
@@ -1,0 +1,10 @@
+import { ProductData, RankingFilter } from '@/types/productType';
+
+export type GetProductRankingRequest = {
+  targetType?: RankingFilter['targetType'];
+  rankType?: RankingFilter['rankType'];
+};
+
+export type GetProductRankingResponse = {
+  products: ProductData[];
+};
