+++ src/services/themeData/index.ts
@@ -1,6 +1,8 @@
 import { BACKEND_API } from '@/constants/api';
 import { ERROR_MESSAGES } from '@/constants/errorMessage';
-import { GetThemesResponse, ThemeHeaderData } from '@/types/themeType';
+import { ThemeHeaderData, ThemeListData } from '@/types/themeType';
+
+import { GetThemesResponse } from './types';
 
 interface FetchThemeHeaderDataResponse {
   themeHeaderContents?: ThemeHeaderData;
@@ -29,5 +31,20 @@ export const fetchThemeHeaderData = async (
       themeHeaderContents: undefined,
       error: ERROR_MESSAGES.FETCH_ERROR,
     };
+  }
+};
+
+interface FetchThemeDataResponse {
+  themes: ThemeListData[];
+  error?: string;
+}
+
+export const fetchThemeData = async (): Promise<FetchThemeDataResponse> => {
+  try {
+    const response = await BACKEND_API.get<GetThemesResponse>('/api/v1/themes');
+
+    return { themes: response.data.themes, error: undefined };
+  } catch (error) {
+    return { themes: [], error: ERROR_MESSAGES.FETCH_ERROR };
   }
 };
