+++ src/services/themeData/types.ts
@@ -1,0 +1,5 @@
+import { ThemeData } from '@/types/themeType';
+
+export type GetThemesResponse = {
+  themes: ThemeData[];
+};
